split classes to separate files made filters inherited
